+++ components/suppliers/DetailsCard.tsx
